requirements.txt
praw
pytesseract
python-dotenv